Fix shift operator issue on fsaa/fsaa.c cppcheck: Shifting signed 32-bit value by 31 bits is undefined behaviour

+++ fsaa/fsaa.c
@@ -48,7 +48,7 @@ unsigned int get_fsaa_mode(const unsigned int index)
 {
 	unsigned int mask;
 
-	mask = 1 << index;
+	mask = ((unsigned int)1) << index;
 
 	if ((fsaa_modes & mask) == mask)
 	{
